Set Referrer Policy
Browsers include a `Referrer` header that identifies where a user came
from when visiting a new page. It's useful in tracking where users are
coming from, but there are some privacy concerns with that. The
`Referrer-Policy` header controls when and how much information is
provided.

- `no-referrer`. Tells the browser to never send the `Referer` header.
- `same-origin`. Send the referrer, but only on requests inside the site
  (e.g. /security-in-the-static-buildpack => /posts)
- `strict-origin`. Send the referrer information to all origins, but
  only the URL sans path (e.g. https://example.com/)
- `strict-origin-when-cross-origin`. Send full referrer information on
  same origin, but only the URL sans path on foreign origin.

`no-referrer` can be used as a fallback for browsers as many of these
options have not yet been implemented at this point.